+++ src/components/core/unread-notification.tsx
@@ -4,15 +4,16 @@ import { FHIRModel } from "@/fhir/models/fhir-model";
 
 export type UnreadNotificationProps = {
   className?: string;
-  data: FHIRModel<fhir4.Resource>[];
+  data: FHIRModel<fhir4.Resource>[] | undefined;
 };
 
 export const UnreadNotification = ({ className, data }: UnreadNotificationProps) => {
   const userBuilderId = useUserBuilderId();
 
-  const unreadOutsideRecords = data.filter(
-    (record) => !record.isDismissed && !record.isRead && !record.ownedByBuilder(userBuilderId)
-  );
+  const unreadOutsideRecords =
+    data?.filter(
+      (record) => !record.isDismissed && !record.isRead && !record.ownedByBuilder(userBuilderId)
+    ) ?? [];
 
   if (unreadOutsideRecords.length > 0) {
     return <UnreadNotificationIcon className={className} />;