+++ src/services/conditions.ts
@@ -12,15 +12,17 @@ import {
 import { getLensBuilderId } from "@/api/urls";
 import { CTWRequestContext } from "@/components/core/providers/ctw-context";
 import { useQueryWithPatient } from "@/components/core/providers/patient-provider";
-import { usePatientBasicResources } from "@/fhir/basic";
 import { ConditionModel } from "@/fhir/models/condition";
-import { useFQSFeatureToggle } from "@/hooks/use-feature-toggle";
 import { createGraphqlClient, fqsRequest } from "@/services/fqs/client";
-import { ConditionGraphqlResponse, conditionsQuery } from "@/services/fqs/queries/conditions";
+import {
+  ConditionGraphqlResponse,
+  conditionsQuery,
+  ConditionWithBasics,
+} from "@/services/fqs/queries/conditions";
 import { cloneDeep, orderBy } from "@/utils/nodash";
 import {
-  QUERY_KEY_OTHER_PROVIDER_CONDITIONS,
   QUERY_KEY_PATIENT_CONDITIONS,
+  QUERY_KEY_PATIENT_SUMMARY_CONDITIONS,
 } from "@/utils/query-keys";
 import { queryClient } from "@/utils/request";
 import { Telemetry, withTimerMetric } from "@/utils/telemetry";
@@ -36,45 +38,31 @@ export function usePatientBuilderConditions(enabled = true) {
 
 function usePatientSummaryConditions(enabled = true) {
   return useQueryWithPatient(
-    QUERY_KEY_OTHER_PROVIDER_CONDITIONS,
+    QUERY_KEY_PATIENT_SUMMARY_CONDITIONS,
     [],
     withTimerMetric(fetchPatientSummaryConditionsFQS, "req.timing.summary_conditions"),
     enabled
   );
 }
 
+// TODO - once Canvas is no longer using the Patient Conditions Outside component, we can remove this
 export function usePatientConditionsOutside() {
   const [conditions, setConditions] = useState<ConditionModel[]>([]);
   const builderConditionsQuery = usePatientBuilderConditions();
   const summaryConditionsQuery = usePatientSummaryConditions();
 
-  // This query is a noop when FQS is disabled and will just return an empty list of basic resources.
-  const basicQuery = usePatientBasicResources();
-
   useEffect(() => {
     const builderConditions = builderConditionsQuery.data ?? [];
-    const basics = basicQuery.data ?? [];
     const outsideConditions = filterOutsideConditions(
       summaryConditionsQuery.data ?? [],
       builderConditions
     );
-    // If basic data came back from the above useBasic call, manually map any basic data to the condition
-    // it corresponds to.
-    if (basics.length > 0) {
-      outsideConditions.forEach((c, i) => {
-        const filteredBasics = basics.filter((b) => b.subject?.reference === `Condition/${c.id}`);
-        outsideConditions[i].revIncludes = filteredBasics;
-      });
-    }
     setConditions(outsideConditions);
-  }, [builderConditionsQuery.data, summaryConditionsQuery.data, basicQuery.data]);
+  }, [builderConditionsQuery.data, summaryConditionsQuery.data]);
 
-  const isLoading =
-    builderConditionsQuery.isLoading || summaryConditionsQuery.isLoading || basicQuery.isLoading;
-  const isError =
-    builderConditionsQuery.isError || summaryConditionsQuery.isError || basicQuery.isError;
-  const isFetching =
-    builderConditionsQuery.isFetching || summaryConditionsQuery.isFetching || basicQuery.isFetching;
+  const isLoading = builderConditionsQuery.isLoading || summaryConditionsQuery.isLoading;
+  const isError = builderConditionsQuery.isError || summaryConditionsQuery.isError;
+  const isFetching = builderConditionsQuery.isFetching || summaryConditionsQuery.isFetching;
 
   return {
     isLoading,
@@ -85,43 +73,24 @@ export function usePatientConditionsOutside() {
 }
 
 export function usePatientConditionsAll(enabled = true) {
-  const fqs = useFQSFeatureToggle("conditions");
   const [conditions, setConditions] = useState<ConditionModel[]>([]);
   const builderConditionsQuery = usePatientBuilderConditions(enabled);
   const summaryConditionsQuery = usePatientSummaryConditions(enabled);
 
-  // This query is a noop when FQS is disabled and will just return an empty list of basic resources.
-  const basicQuery = usePatientBasicResources();
-
   useEffect(() => {
     const builderConditions = builderConditionsQuery.data ?? [];
-    const basics = basicQuery.data ?? [];
     const allConditions = filterOutsideConditions(
       summaryConditionsQuery.data ?? [],
       builderConditions
     );
     allConditions.push(...builderConditions);
 
-    // If basic data came back from the above useBasic call, manually map any basic data to the condition
-    // it corresponds to.
-    if (basics.length > 0) {
-      allConditions.forEach((c, i) => {
-        const filteredBasics = basics.filter((b) => b.subject?.reference === `Condition/${c.id}`);
-        allConditions[i].revIncludes = filteredBasics;
-      });
-    }
     setConditions(allConditions);
-  }, [builderConditionsQuery.data, summaryConditionsQuery.data, basicQuery.data]);
+  }, [builderConditionsQuery.data, summaryConditionsQuery.data]);
 
-  const isLoading =
-    builderConditionsQuery.isLoading || summaryConditionsQuery.isLoading || basicQuery.isLoading;
-  const isError =
-    builderConditionsQuery.isError || summaryConditionsQuery.isError || basicQuery.isError;
-  const isFetching =
-    builderConditionsQuery.isFetching ||
-    summaryConditionsQuery.isFetching ||
-    basicQuery.isFetching ||
-    !fqs.ready;
+  const isLoading = builderConditionsQuery.isLoading || summaryConditionsQuery.isLoading;
+  const isError = builderConditionsQuery.isError || summaryConditionsQuery.isError;
+  const isFetching = builderConditionsQuery.isFetching || summaryConditionsQuery.isFetching;
 
   return {
     isLoading,
@@ -131,8 +100,8 @@ export function usePatientConditionsAll(enabled = true) {
   };
 }
 
-function setupConditionModelsWithFQS(conditionResources: fhir4.Condition[]): ConditionModel[] {
-  return conditionResources.map((c) => new ConditionModel(c));
+function setupConditionModelsWithFQS(conditionResources: ConditionWithBasics[]): ConditionModel[] {
+  return conditionResources.map((c) => new ConditionModel(c, undefined, c.BasicList));
 }
 
 function filterAndSort(conditions: ConditionModel[]): ConditionModel[] {